add css and js folder and finish the UserDao

+++ src/main/java/com/photomap/web/dao/impl/UserMapper.java
@@ -8,7 +8,7 @@ import com.photomap.web.model.User;
 public class UserMapper implements RowMapper<User> {
 	public User mapRow(ResultSet rs, int rowNum) throws SQLException {
 		User user = new User();
-		user.setId(rs.getLong("id"));
+		user.setId(rs.getInt("id"));
 		user.setUsername(rs.getString("username"));
 		user.setPassword(rs.getString("password"));
 		user.setEmail(rs.getString("email"));
